Data.Graph.Dag
practicing on a simpler case.

a free symmetric premonoidal category is an open DAG (with some incoming
edges before the first node and some outgoing edges after the last
node), in which each node is labelled by an arrow in the quiver which
determines the nodes's in- and out-degrees, and in which every edge is
typed.

a DAG is just a DAG.